dfs looking mostly good
the problem is this: when i have a state block adjacent to transition blocks, we're good. when i add another state block, the second state block also acts as the parent of the same transition blocks. when i delete either of the two state blocks, the transition blocks are freed.